Create helpers function, fix date range validations
